Use ints for the biome surface block
Fixes the surface block being -97 instead of 159 in the mesa biome.

+++ common/src/main/java/com/khorn/terraincontrol/LocalBiome.java
@@ -83,14 +83,14 @@ public interface LocalBiome
      * 
      * @return The default SurfaceBlock.
      */
-    byte getSurfaceBlock();
+    int getSurfaceBlock();
 
     /**
      * Gets the default value for the GroundBlock setting.
      * 
      * @return The default GroundBlock.
      */
-    byte getGroundBlock();
+    int getGroundBlock();
 
     /**
      * Gets the temperature at the given position, if this biome would be
